Create README.md
Created an application that is used to keep a record of employees with details. Users need to create or signup users then login to the system and add or list or update employees with employee details on the dashboard. For storing data we used MongoDB as a database using REST API calls. For security, we have Used the JWT token to encrypt the password and…